Update main.cpp
The implementation in question is a "movie data store", applying studies on Binary Search Trees. First, the tree is loaded with some objects (created altogether), and from there, the application becomes interactive, greeting the user with "Hello, welcome!", asking if he "wants to start? (1-YES, 2-NO)", and starting the do-while loop if option 1 is selected (if it is not , the program ends). Then, the program asks "And now, what do you want?" and the options below are shown:
1 - Know if this tree contains films.
2 - Know the oldest film of this tree.
3 - Know the most recent film of this tree.
4 - Know if the requested film is registered in this tree.
5 - Register a movie in this tree.
6 - Remove a film from this tree.
7 - View all the films registered in this tree.
8 - Clean the tree.
Option 9 and its associated commands refer to cleaning the display screen. As Netbeans has its own terminal (which does not recognize the command system (" cls ") or others "system"), I chose to just comment.


Note: the first part of the implementation is just a test for the thought logic, and that is why it was commented. Consisting of:
test 1: transform the characters for the day into a whole number and print it;
test 2: check the date comparison method. if one date is less than the other, print "it worked", if it is not, print "it went wrong";
test 3: verification of the try-catch condition defined in the class "Datas.h". If the date is created successfully, print "excellent", if it is not, print "exception".